Create 3d_calcium_wave_contouring.py
Code to generate a figure for the NEURON 3D Methods paper, (here as Figure 1A of figure1.pdf), showing the propagation over time of a calcium wave on a 3D model of a neuron. 
Ultimately, it is a great visualization tool for seeing how calcium concentration changes in a cell, and how it can be tracked. The work around mesh grids and 3D image flattening was the most challenging here. A lot of the int's in this file are specific to the cell model used, for best visual results, but are ultimately arbitrary.