Add files via upload
BankCLI/BankGUI are from an Object Oriented Programming class first assignment. It simulates a very basic bank-type program.

HanaG_FOL_Pset3_input/output.txt are my submissions to our 'Software Analysis and Verification' course, where we had to use dafny to "Prove that the length of the concatenation of two strings is equal to the sum of the lengths of the individual strings." using First Order Logic and theorem proving programs.

Psched.c is from my Data Structures & Programming course in freshman year where the assignment was to (according to specifications) make a program for program-scheduling, given input heuristics and preferences.